+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+
+#TODO: Sort the following line into the file
+- ClassEndsWithBlankLineBeforeClosingBrace rule (braces) - Require a newline before the closing brace of a class
+
 # CodeNarc Change Log
 
 Version 1.2.1 (Aug 2018)